Task2/Backend/requirements.txt
flask
flask_sqlalchemy
flask-cors
gunicorn
langchain-google-genai
langchain-core
langchain
google-genai
pandas
numpy
python-dotenv